refactor: extract settings-validation from ipc-handlers
